Domain independent update
This version allows for multiple domains to be active at the same time, users from domain a are directed towards domain a.

This works as long as a user is a member of only one domain.

Next to this a lot of static delays have been removed to now accomodate more flexible asynchronous functions.
Effectively this lowers the loading time of the pages while still being sufficient for loading essential elements like the domainjson